pyproject.toml
[tool.poetry]
name = "superBOM"
version = "0.1.0"
description = ""
authors = ["Michael Beale <michael.beale@intel.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "superbom", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
pandas = "^2.2.3"
PyYAML = "^6.0.2"
tqdm = "^4.67.0"
spdx-matcher = "^0.0.13"
spdx-lookup = "^0.3.3"
openpyxl = "^3.1.5"
foss-flame = "^0.20.7"
colorlog = "^6.9.0"
poetry-core = "^1.9.1"
packaging = "^24.2"

[tool.poetry.scripts]
superbom = "superbom.main:main"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
